test: update valid.csv with additional country data and enhance plugin tests
