+++ ckanext/validate/tests/files_test/valid.csv
@@ -1,4 +1,8 @@
-id,name,price
-1,Alpha,10
-2,Beta,20
-3,Gamma,30
+"PAIS","ANIO_DESEMBOLSO","DESCRIPCION","PRODUCTO","FONDO","MODALIDAD","MONTO_BRUTO_USD"
+Costa Rica,"2012",PRE-11/2012 - Costa Rica (CRI) - FONTEC - Fondo de Cooperación Técnica - No Reembolsable - 302208,Cooperación Técnica,Fondos Ordinarios,No Reembolsable,"10000.00"
+El Salvador,"2011",DI-120/2011 - El Salvador (SLV) - Fondos Ordinarios - No Reembolsable - 302194,Ayuda de Emergencia y por Desastres Naturales,Fondos Ordinarios,No Reembolsable,"250000.00"
+Honduras,"2011",DI-124/2011 - Honduras (HND) - Fondos Ordinarios - No Reembolsable - 302196,Ayuda de Emergencia y por Desastres Naturales,Fondos Ordinarios,No Reembolsable,"250000.00"
+Guatemala,"2011",DI-11/2011 - Guatemala (GTM) - Fondos Ordinarios - No Reembolsable - 302130,Cooperación de Países Fundadores y Regionales no Fundadores,Fondos Ordinarios,No Reembolsable,"500000.00"
+Cuba,"2020",DI-137/2020 - Cuba (CUB) - Fondos Ordinarios - No Reembolsable - 500564,Ayuda de Emergencia y por Desastres Naturales,Fondos Ordinarios,No Reembolsable,"1000000.00"
+Colombia,"2017",DI-29/2017 - Colombia (COL) - Fondos Ordinarios - No Reembolsable - 500275,Ayuda de Emergencia y por Desastres Naturales,Fondos Ordinarios,No Reembolsable,"250000.00"
+Argentina,"2013",DI-38/2013 - Argentina (ARG) - Fondos Ordinarios - No Reembolsable - 302348,Ayuda de Emergencia y por Desastres Naturales,Fondos Ordinarios,No Reembolsable,"249804.64"
