+++ ckanext/validate/tests/test_plugin.py
@@ -1,6 +1,4 @@
 """
-Tests for plugin.py.
-
 Tests are written using the pytest library (https://docs.pytest.org), and you
 should read the testing guidelines in the CKAN docs:
 https://docs.ckan.org/en/2.9/contributing/testing.html
@@ -49,9 +47,43 @@ To temporary patch the CKAN configuration for the duration of a test you can use
 """
 import pytest
 from ckan.plugins import plugin_loaded
+from ckanext.validate.actions import action as validate_action
+from ckanext.validate.auth import validation as validate_auth
+from ckanext.validate.blueprints import resource as validate_resource
+from ckanext.validate.plugin import ValidatePlugin
 
 
 @pytest.mark.ckan_config("ckan.plugins", "validate")
 @pytest.mark.usefixtures("with_plugins")
 def test_plugin():
     assert plugin_loaded("validate")
+
+
+def test_plugin_registers_expected_actions():
+    plugin = ValidatePlugin()
+
+    actions = plugin.get_actions()
+
+    assert actions == {
+        "resource_validate": validate_action.resource_validate,
+        "resource_validation_show": validate_action.resource_validation_show,
+    }
+
+
+def test_plugin_registers_expected_auth_functions():
+    plugin = ValidatePlugin()
+
+    auth_functions = plugin.get_auth_functions()
+
+    assert auth_functions == {
+        "resource_validate": validate_auth.resource_validate,
+        "resource_validation_show": validate_auth.resource_validation_show,
+    }
+
+
+def test_plugin_registers_expected_blueprint():
+    plugin = ValidatePlugin()
+
+    blueprints = plugin.get_blueprint()
+
+    assert blueprints == [validate_resource.resource_validate_blueprint]
